sanxei95 - * refactor spageti code
