+++ src/features/views/components/view-detail/view-detail.tsx
@@ -1,203 +1,48 @@
+import { useEffect, useState } from "react";
 import { ArrowLeft, Edit, Trophy, Loader2 } from "lucide-react";
-import { useAppSelector } from "@/app/hooks.ts";
-import "./view-detail.css";
-import { useMemo, useRef, useState } from "react";
 import { useParams, useNavigate, useLocation } from "react-router-dom";
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import { RaiderioProfile } from "@/features/views/api/raiderio.ts";
-import { View } from "@/features/views/model/view.ts";
-import { selectUsername } from "@/app/authSlice.ts";
-import { userRequestVoid } from "@/shared/api/httpClient.ts";
-import { ViewRequest } from "@/features/views/api/view-types.ts";
-import { haveSameCharacters } from "@/features/views/utils.ts";
+import { useViewDetail } from "@/features/views/hooks/useViewDetail.ts";
 import { CharacterLadder } from "./character-ladder/character-ladder.tsx";
 import { DungeonGrid } from "./dungeon-grid/dungeon-grid.tsx";
 import { EditView } from "./actions/edit-view.tsx";
 import { SyncErrorDialog } from "./actions/sync-error-dialog.tsx";
-import {
-  viewKeys,
-  fetchViewData,
-  fetchCachedViewData,
-  fetchWowStatic,
-} from "@/features/views/api/viewQueries.ts";
-
-interface ViewEditMeta {
-  pendingCharacters: RaiderioProfile[];
-}
-
-const UUID_REGEX =
-  /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
+import "./view-detail.css";
 
 export function ViewDetail() {
   const { viewId } = useParams();
   const navigate = useNavigate();
-  const onBack = () => navigate("/");
   const location = useLocation();
-  const queryClient = useQueryClient();
-  const username = useAppSelector(selectUsername);
 
   const locationState = location.state as { owner?: string; entitiesCount?: number } | null;
   const owner = locationState?.owner ?? null;
   const entitiesCount = locationState?.entitiesCount ?? 0;
-  const canEdit = username !== null && username === owner;
 
   const [isEditOpen, setIsEditOpen] = useState(false);
-  const [expectedCount, setExpectedCount] = useState(entitiesCount);
-  const editPollCountRef = useRef(0);
-  const lastSeenDataUpdatedAt = useRef(0);
 
-  const isViewIdValid = !!viewId && UUID_REGEX.test(viewId);
-  const safeViewId = viewId ?? "";
+  const {
+    profiles,
+    cachedProfiles,
+    viewName,
+    season,
+    initialized,
+    editMeta,
+    syncError,
+    canEdit,
+    isViewIdValid,
+    expectedCount,
+    saveCharacters,
+    clearSyncError,
+  } = useViewDetail(viewId, owner, entitiesCount);
 
-  if (viewId && !UUID_REGEX.test(viewId)) {
-    navigate("/");
-  }
 
-  const { data: rawViewData, isLoading, dataUpdatedAt } = useQuery({
-    queryKey: viewKeys.data(safeViewId),
-    queryFn: () => fetchViewData(safeViewId),
-    enabled: isViewIdValid,
-    refetchInterval: (query) => {
-      const hasPendingScore = query.state.data?.data.some((c) => c.score === -1);
-      const hasEditMeta = !!queryClient.getQueryData<ViewEditMeta>(
-        viewKeys.editMeta(safeViewId),
-      );
-      const isSyncingInitial =
-        !hasEditMeta && expectedCount > 0 && (query.state.data?.data.length ?? 0) === 0;
-      return hasPendingScore || hasEditMeta || isSyncingInitial ? 3000 : false;
-    },
-  });
-
-  const { data: cachedData } = useQuery({
-    queryKey: viewKeys.cachedData(safeViewId),
-    queryFn: () => fetchCachedViewData(safeViewId),
-    enabled: isViewIdValid,
-    staleTime: Infinity,
-  });
-
-  const { data: season } = useQuery({
-    queryKey: viewKeys.static(),
-    queryFn: fetchWowStatic,
-    staleTime: Infinity,
-  });
-
-  const { data: editMeta } = useQuery({
-    queryKey: viewKeys.editMeta(safeViewId),
-    queryFn: () => null as ViewEditMeta | null,
-    enabled: false,
-    staleTime: Infinity,
-    gcTime: 5 * 60 * 1000,
-    initialData: null,
-  });
-
-  const { data: syncError } = useQuery({
-    queryKey: viewKeys.syncError(safeViewId),
-    queryFn: () => null as RaiderioProfile[] | null,
-    enabled: false,
-    staleTime: Infinity,
-    gcTime: 5 * 60 * 1000,
-    initialData: null,
-  });
-
-  const MAX_EDIT_POLLS = 5;
-
-  const profiles = useMemo(() => {
-    const apiData = rawViewData?.data ?? [];
-    if (!editMeta) return apiData;
-
-    const { pendingCharacters } = editMeta;
-
-    const key = (c: RaiderioProfile) => c.name.toLowerCase();
-
-    const apiByKey = new Map(apiData.map((c) => [key(c), c]));
-    const pendingKeys = new Set(pendingCharacters.map(key));
-
-    const backendCaughtUp =
-      apiData.every((c) => pendingKeys.has(key(c))) &&
-      pendingCharacters.every((c) => apiByKey.has(key(c)));
-
-    if (backendCaughtUp) {
-      editPollCountRef.current = 0;
-      queryClient.setQueryData(viewKeys.editMeta(safeViewId), null);
-      return apiData;
+  useEffect(() => {
+    if (viewId && !isViewIdValid) {
+      navigate("/");
     }
+  }, [viewId, isViewIdValid, navigate]);
 
-    // Count each new poll (dataUpdatedAt increases on every successful refetch)
-    if (dataUpdatedAt > lastSeenDataUpdatedAt.current) {
-      lastSeenDataUpdatedAt.current = dataUpdatedAt;
-      editPollCountRef.current += 1;
-      if (editPollCountRef.current >= MAX_EDIT_POLLS) {
-        editPollCountRef.current = 0;
-        const failed = pendingCharacters.filter((c) => !apiByKey.has(key(c)));
-        if (failed.length > 0) {
-          queryClient.setQueryData(viewKeys.syncError(safeViewId), failed);
-        }
-        queryClient.setQueryData(viewKeys.editMeta(safeViewId), null);
-        return apiData;
-      }
-    }
-
-    return pendingCharacters.map((c) => apiByKey.get(key(c)) ?? c);
-  }, [rawViewData, editMeta, safeViewId, queryClient, dataUpdatedAt]);
-
-  const cachedProfiles = cachedData?.data ?? [];
-  const viewName = rawViewData?.viewName ?? "";
-  const initialized = !isLoading;
-
-  const saveMutation = useMutation({
-    mutationFn: (characters: RaiderioProfile[]) => {
-      const request: ViewRequest = {
-        name: viewName,
-        entities: characters.map((c) => ({
-          name: c.name,
-          region: c.region,
-          realm: c.realm,
-          type: "com.kos.entities.domain.WowEntityRequest",
-        })),
-        published: true,
-        featured: false,
-        game: "WOW",
-      };
-      return userRequestVoid("PUT", `/views/${viewId}`, request);
-    },
-
-    onMutate: async (characters) => {
-      editPollCountRef.current = 0;
-      lastSeenDataUpdatedAt.current = dataUpdatedAt;
-      queryClient.setQueryData(viewKeys.syncError(safeViewId), null);
-      setExpectedCount(characters.length);
-      await queryClient.cancelQueries({ queryKey: viewKeys.data(safeViewId) });
-      await queryClient.cancelQueries({ queryKey: viewKeys.ownList() });
-
-      queryClient.setQueryData<ViewEditMeta>(viewKeys.editMeta(safeViewId), {
-        pendingCharacters: characters,
-      });
-
-      queryClient.setQueryData<View[]>(viewKeys.ownList(), (old) =>
-        old?.map((v) =>
-          v.id === viewId
-            ? {
-                ...v,
-                simpleView: {
-                  ...v.simpleView,
-                  entitiesIds: characters.map((_, i) => i),
-                },
-              }
-            : v,
-        ) ?? [],
-      );
-    },
-
-    onError: () => {
-      queryClient.setQueryData(viewKeys.editMeta(safeViewId), null);
-      queryClient.invalidateQueries({ queryKey: viewKeys.ownList() });
-    },
-  });
-
-  const handleSavedCharacters = async (characters: RaiderioProfile[]) => {
-    if (!haveSameCharacters(characters, profiles)) {
-      await saveMutation.mutateAsync(characters);
-    }
+  const handleSavedCharacters = async (characters: typeof profiles) => {
+    await saveCharacters(characters);
     setIsEditOpen(false);
   };
 
@@ -207,7 +52,7 @@ export function ViewDetail() {
     <div className="view-detail-container">
       <div className="view-detail-content">
         <div className="view-detail-header">
-          <button onClick={onBack} className="header-back-button">
+          <button onClick={() => navigate("/")} className="header-back-button">
             <ArrowLeft className="header-icon" />
           </button>
           <h1 className="header-view-title">{viewName}</h1>
@@ -224,7 +69,7 @@ export function ViewDetail() {
           )}
         </div>
 
-        {!isLoading && profiles.length === 0 && expectedCount > 0 ? (
+        {profiles.length === 0 && expectedCount > 0 ? (
           <div className="syncing-state">
             <Loader2 className="syncing-icon" />
             <h3 className="syncing-title">Syncing characters…</h3>
@@ -232,7 +77,7 @@ export function ViewDetail() {
               Your characters are being prepared. This usually takes a few seconds.
             </p>
           </div>
-        ) : !isLoading && profiles.length === 0 ? (
+        ) : profiles.length === 0 ? (
           <div className="empty-state">
             <Trophy className="empty-icon" />
             <h3 className="empty-title">No characters in this view</h3>
@@ -240,10 +85,7 @@ export function ViewDetail() {
               Add characters to start tracking their Mythic+ progress
             </p>
             {canEdit && (
-              <button
-                className="empty-add-btn"
-                onClick={() => setIsEditOpen(true)}
-              >
+              <button className="empty-add-btn" onClick={() => setIsEditOpen(true)}>
                 Add Your First Character
               </button>
             )}
@@ -253,9 +95,8 @@ export function ViewDetail() {
             <CharacterLadder
               characters={profiles}
               cachedCharacters={cachedProfiles}
-              season={season ?? null}
+              season={season}
             />
-
             {season && (
               <DungeonGrid
                 raiderioProfiles={profiles}
@@ -276,9 +117,8 @@ export function ViewDetail() {
 
       <SyncErrorDialog
         failedCharacters={syncError ?? []}
-        onClose={() => queryClient.setQueryData(viewKeys.syncError(safeViewId), null)}
+        onClose={clearSyncError}
       />
-
     </div>
   );
 }
